Updated inference.py for text generation task
