+++ SupportIQ/src/inference_generation.py
@@ -1,0 +1,60 @@
+from transformers import AutoTokenizer, AutoModelForSeq2SeqLM
+import torch
+import os
+import boto3
+
+def model_fn():
+
+    local_dir = "/app/model/"
+
+    if not os.path.exists(local_dir):
+        os.makedirs(local_dir)
+    
+    # s3 = boto3.client('s3')
+    # objects = s3.list_objects_v2(Bucket=s3_bucket, Prefix=s3_prefix)
+    
+    # for obj in objects.get("Contents", []):
+    #     key = obj['Key']
+    #     if key.endswith("/"):
+    #         continue
+    #     file_name = os.path.basename(key)
+    #     local_path = os.path.join(local_dir, file_name)
+    #     s3.download_file(s3_bucket, key, local_path)
+
+    tokenizer = AutoTokenizer.from_pretrained(local_dir)                          #load tokenizer
+    
+    model = AutoModelForSeq2SeqLM.from_pretrained(local_dir)         #load model
+    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+    model.to(device)
+    model.eval()
+
+    return {'model': model,'tokenizer': tokenizer, 'device':device}
+
+
+def predict_fn(input_text, model_obj):
+
+    if input_text is None:
+        raise ValueError("Inpute text is required under key 'input'")
+    
+    tokenizer = model_obj['tokenizer']
+    model = model_obj['model']
+    device = model_obj['device']
+
+    if isinstance(input_text,list):
+        inputs = tokenizer(input_text, return_tensors='pt',padding=True, truncation=True).to(device)
+    else:
+        inputs = tokenizer([input_text], return_tensors='pt',padding=True, truncation=True).to(device)
+
+    with torch.no_grad():
+        outputs = model.generate(**inputs, max_length=128,
+                                early_stopping=True)    
+        predicted_ids = tokenizer.decode(outputs[0], skip_special_tokens=True)
+    
+    return predicted_ids
+
+def predict(input_text):
+    model_obj = model_fn()
+    return predict_fn(input_text, model_obj)
+
+
+
